# [Core] Optimize ProcessWrapper threads usage.