This moves the ProcessWrapper to use tasks instead of threads. The net benefit of this is that we use 60 less threads for loading Main.sln, as the threads are now assigned to threadpool threads.
Reduces total time spent in CaptureError from 60s to 0.8ms and CaptureOutput from 64s to 33ms.