@Therzok 2016-09-26
Maybe we should go further and only create one task and do `Task.WhenAny` internally?