@alanmcgovern 2016-09-26
It seems fine to me. I don't think there's any real benefit to trying to reduce the tasks further.